refactor(frontend): lift shared timer types/api/format into app/frontend/lib/timers
Phase 6c (timesheet) reuses the same wire types, API client, and
time/date helpers that 6b's timers-calendar shipped, so they belong
in a shared lib instead of being duplicated per island.

- Move `types.ts` and `api.ts` to `app/frontend/lib/timers/`.
- Split `format.ts`: generic time helpers (parseHHMM, formatHHMM,
  runningDuration, isStartable, todayISO, addDays, weekRange,
  monthRange, …) move to `lib/timers/format.ts`. Calendar-grid-
  specific bits (`buildWeeks`, `monthLabel`, `DayCell`, `Week`)
  stay in the timers-calendar island as `calendar.ts`.
- Update timers-calendar imports to the new paths.

No behavior change.
