+++ app/frontend/islands/timers-calendar/components/TimerBadge.vue
@@ -5,8 +5,8 @@
 // `app/assets/stylesheets/partials/_calendar.scss`.
 
 import {computed, onBeforeUnmount, onMounted, ref} from "vue"
-import {formatHHMM, runningDuration} from "../format"
-import type {Timer} from "../types"
+import {formatHHMM, runningDuration} from "../../../lib/timers/format"
+import type {Timer} from "../../../lib/timers/types"
 
 const props = defineProps<{timer: Timer}>()
 defineEmits<{click: []}>()
